gestion des categories, CQRS et ajout d'exception

+++ src/catalog/deleteCategory/deleteCategory.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Category } from '../category.entity';
@@ -11,6 +11,13 @@ export class DeleteCategoryService {
   ) {}
 
   async execute(id: string): Promise<void> {
+
+    const category = await this.categoryRepository.findOne({ where: { id } });
+
+    if (!category) {
+      throw new NotFoundException('Category not found');
+    }
+    
     await this.categoryRepository.delete(id);
   }
 }
